reunion  and   demande  action  test

+++ src/app/modules/action-simplifier/actionaction-simplifier.component.ts
@@ -12,6 +12,10 @@ import { SpinirLoadService } from 'src/app/shared/spinir-load.service';
 import { Router } from '@angular/router';
 import { ActSimplifierDetailsComponent } from './act-simplifier-details/act-simplifier-details.component';
 import { ActSimplifierGeniriqueComponent } from './act-simplifier-genirique/act-simplifier-genirique.component';
+import { MessageAlertService } from 'src/app/shared/message-alert.service';
+import { TokenStorageService } from 'src/app/authentification/_services/token-storage.service';
+import { UserService } from 'src/app/authentification/_services/user.service';
+import { ConfirmationDialogComponentComponent } from 'src/app/shared/confirmation-dialog-component/confirmation-dialog-component.component';
 
 @Component({
   selector: 'app-actionaction-simplifier',
@@ -24,47 +28,81 @@ export class ActionactionSimplifierComponent {
 
 
  
-  displayedColumns = ['id', 'Déchlencheur','filiale.D', 'Responsable traitement', 'filiale.R','Responsable Côlture','filiale.C','Responsable suivi','filiale.S','...','....'];
+  displayedColumns = ['id', 'Déchlencheur','filiale.D', 'Responsable traitement', 'filiale.R','Responsable Côlture','filiale.C','...'];
   dataSource!: MatTableDataSource<Action>;
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
   
+  
   isLoading = true;
+  role!:boolean;
 
-  constructor(private router :Router ,public dialog: MatDialog,private serviceAction: ActionServiceService,spinnerService :SpinirLoadService) {
-    // Create 100 users
+  isLoading2=false;
+  currentUser: any;
+
+  constructor(private router: Router, private toast :MessageAlertService,
+    private token: TokenStorageService,
+    private userService:UserService,
+    public dialog: MatDialog,
+    private serviceAction: ActionServiceService,
+    spinnerService: SpinirLoadService) {
+      console.log(this.token.getUser)
+      this.currentUser = this.token.getUser();
+  this.userService.getEmployee(this.currentUser.id).subscribe((res)=>{
+    this.currentUser=res;
+  
+  });
+  console.log(this.currentUser);
+  this.role=this.currentUser.roles[0]=="ADMIN";
     this.loadData()
-}
+  }
 
-    // Assign the data to the data source for the table to render
+  // Assign the data to the data source for the table to render
 
-loadData(){
-  let actions: Action[] = [];
-let  id:number=1;
-  this.serviceAction.getActionAllAction(id).subscribe(res=>{
-    this.isLoading=false;
-   actions=res;
-   this.dataSource = new MatTableDataSource(actions);
-   this.dataSource.paginator = this.paginator;
-   this.dataSource.sort = this.sort;
-  }, 
-  error => this.isLoading = false
-);
-}
+  loadData() {
+    let actions: Action[] = [];
+    let id: number = 1;
+    this.serviceAction.getActionAllAction(id).subscribe(res => {
+      this.isLoading = false;
+      actions = res;
+      this.dataSource = new MatTableDataSource(actions);
+      this.dataSource.paginator = this.paginator;
+      this.dataSource.sort = this.sort;
+    },
+      error => this.isLoading = false
+    );
+  }
 
 runSenarioSimplifier(idScenario:number):void{
+  this.isLoading2 = true;
   console.log(idScenario)
+  this.toast.messageSuccess("Scénario prêt")
   this.serviceAction.runSenarioSimplifier(idScenario).subscribe(
-    res=>{
-      console.log(res)
+    (res) => {
+      console.log(res);
+     
+    if(res=="ok"){
+      
+      
+      this.isLoading2 = false
     }
-  );
-
+  
+      
+    },
+      (error) => this.isLoading2 = false
+    );
 }
 closeScenario(idScenario:number):void{
  
 
+  console.log(idScenario)
+  this.serviceAction.closeScenario(idScenario).subscribe(
+    res => {
+      console.log(res)
+      this.toast.messageErro("Scénario arrét")
+    }
+  );
 
 }
 
@@ -75,7 +113,7 @@ closeScenario(idScenario:number):void{
 
 
   openDialog(): void {
-    this.router.navigate(['scsepcifique']); // Navigate to the 'other' route
+    this.router.navigate(['scsepcifiqueActSimplifer']); // Navigate to the 'other' route
 
    /* const dialogRef = this.dialog.open(AddSecActionComponent, {
       width: '90%', height: '90%', disableClose: true
@@ -84,7 +122,7 @@ closeScenario(idScenario:number):void{
 
   openDialog2(): void {
     const dialogRef = this.dialog.open(ActSimplifierGeniriqueComponent, {
-      width: '60%', height: '90%', disableClose: true
+      width: '85%', height: '90%', disableClose: true
     });
     dialogRef.afterClosed().subscribe((res)=>{
       if(res)
@@ -92,17 +130,34 @@ closeScenario(idScenario:number):void{
     })
   }
 deleteSecenario(idScenario:number){
-  this.serviceAction.deleteAction(idScenario).subscribe(
-    res=>{console.log(res)
-    this.loadData();
-    },error=>{
-      console.log(error)
+  const dialogRef = this.dialog.open(ConfirmationDialogComponentComponent
+    , {
+    width: '300px',
+    data: {
+      title: 'Confirm Deletion',
+      message: 'Are you sure you want to delete this item?'
     }
-  )
+  });
+
+  dialogRef.afterClosed().subscribe(result => {
+    if (result === true) {
+    
+        this.serviceAction.deleteAction(idScenario).subscribe(
+          res => {
+            console.log(res)
+            this.loadData();
+          }, error => {
+            console.log(error)
+          }
+        )
+      
+    }
+  });
 }
+
 openDialogDetails(idScenario:number):void{
   const dialogRef = this.dialog.open(ActSimplifierDetailsComponent, {
-    width: '80%',
+    minWidth: '100%',maxWidth:'100%',width:'100%', height: '100%', disableClose: true,
     data: { idScenario: idScenario} // Pass the parameters as an object
   });
